new commit

+++ src/OsonCommerce.Application/Features/Cashbox/Commands/Delete/DeleteCashboxCommandHandler.cs
@@ -2,27 +2,29 @@ using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 using OsonCommerce.Domain.Entities;
-using FluentValidation;
 using OsonCommerce.Application.Interfaces;
+using OsonCommerce.Application.Exceptions;
 
 namespace OsonCommerce.Application.Features
 {
     public class DeleteCashboxCommandHandler : IRequestHandler<DeleteCashboxCommand>
     {
         private readonly IRepository<Cashbox> _repository;
-        private readonly IValidator<DeleteCashboxCommand> _validator;
         private readonly IUnitOfWork _unitOfWork;
 
-        public DeleteCashboxCommandHandler(IRepository<Cashbox> repository, IValidator<DeleteCashboxCommand> validator, IUnitOfWork unitOfWork)
+        public DeleteCashboxCommandHandler(IRepository<Cashbox> repository, IUnitOfWork unitOfWork)
         {
             _repository = repository;
-            _validator = validator;
             _unitOfWork = unitOfWork;
         }
 
         public async Task Handle(DeleteCashboxCommand request, CancellationToken cancellationToken)
         {
-            await _validator.ValidateAsync(request, cancellationToken);
+            if (request.Id == Guid.Empty)
+            {
+                throw new EmptyRequestException("Id is required");
+            }
+
             await _repository.DeleteAsync(request.Id, cancellationToken);
             await _unitOfWork.SaveChangesAsync(cancellationToken);
         }
