+++ src/OsonCommerce.Application/Features/Category/Commands/Delete/DeleteCategoryCommandHandler.cs
@@ -1,7 +1,7 @@
 using OsonCommerce.Domain.Entities;
 using MediatR;
-using FluentValidation;
 using OsonCommerce.Application.Interfaces;
+using OsonCommerce.Application.Exceptions;
 
 
 namespace OsonCommerce.Application.Features;
@@ -10,18 +10,20 @@ public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryComman
 {
     private readonly IRepository<Category> _repository;
     private readonly IUnitOfWork _unitOfWork;
-    private readonly IValidator<DeleteCategoryCommand> _validator;
 
-    public DeleteCategoryCommandHandler(IRepository<Category> repository, IUnitOfWork unitOfWork, IValidator<DeleteCategoryCommand> validator)
+    public DeleteCategoryCommandHandler(IRepository<Category> repository, IUnitOfWork unitOfWork)
     {
         _repository = repository;
         _unitOfWork = unitOfWork;
-        _validator = validator;
     }
 
     public async Task Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
     {
-        await _validator.ValidateAndThrowAsync(request, cancellationToken);
+        if (request.Id == Guid.Empty)
+        {
+            throw new EmptyRequestException("Id is required");
+        }
+
         await _repository.DeleteAsync(request.Id, cancellationToken);
         await _unitOfWork.SaveChangesAsync(cancellationToken);
     }
